fix: MLX softmax in float32 to match torch.autocast policy
torch.autocast runs softmax in float32 by policy. Global attention
blocks in stage 3 process 16384 tokens at 2048 resolution. fp16
softmax over that many tokens loses precision in the attention
distribution, causing blocky alpha edges.

Previous attempt crashed because norm WEIGHTS were also float32,
causing type promotion cascade. Now norm weights stay fp16 (inputs
temporarily upcast), so only the attention scores allocate fp32.

+++ CorridorKeyModule/core/mlx_model.py
@@ -124,8 +124,10 @@ class MaskUnitAttention(nn.Module):
             q = mx.max(q, axis=3)  # [B, heads, num_win, tokens/stride, head_dim]
 
         # Scaled dot-product attention
+        # Softmax in float32 to match torch.autocast policy (fp16 softmax
+        # loses precision over 16K+ token sequences in global attention)
         attn = (q * self.scale) @ k.transpose(0, 1, 2, 4, 3)  # [..., Nq, Nk]
-        attn = mx.softmax(attn, axis=-1)
+        attn = mx.softmax(attn.astype(mx.float32), axis=-1).astype(q.dtype)
         x = attn @ v  # [B, heads, num_win, Nq, head_dim]
 
         # Reshape back: [B, heads, num_win, Nq, head_dim] -> [B, N_out, dim_out]
